introduce AbstractCollectorRector to avoid mixing AbstractRector and interface
